🚀 AI Brother v2.0.0 - Fully Functional Release
✅ Features:
- Real file upload system (PDFs, docs, images)
- Working camera integration with multiple modes
- Complete image gallery with OCR simulation
- Modern Material 3 UI with navigation
- File management and storage analytics

🔧 Technical:
- Fixed all build errors and Kotlin issues
- 6.0MB APK successfully built
- Supports Android 5.0+ (API 21-28)
- ARM64 and ARMv7 architecture support

📱 Ready for testing and installation!

+++ src/main/java/com/prelightwight/aibrother/FilesFragment.kt
@@ -618,7 +618,7 @@ class FilesFragment : Fragment() {
         var size: Long = 0
         val inputStream: InputStream? = requireContext().contentResolver.openInputStream(uri)
         inputStream?.use {
-            size = it.available()
+            size = it.available().toLong()
         }
         return size
     }
